Implement integration tests for Cliente and Movimiento services; enhance CuentaService with transaction management and exception handling; refactor models to use BigDecimal for monetary values; update DTOs for validation; improve global exception handling; add report generation functionality.

+++ ms-cuentas/src/main/java/com/banco/mscuentas/dto/CuentaResponseDTO.java
@@ -2,6 +2,8 @@ package com.banco.mscuentas.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,8 +14,8 @@ public class CuentaResponseDTO {
     private Long id;
     private String numeroCuenta;
     private String tipoCuenta;
-    private Double saldoInicial;
-    private Double saldoDisponible;
+    private BigDecimal saldoInicial;
+    private BigDecimal saldoDisponible;
     private Boolean estado;
     private String clienteId;
 }
