+++ ms-cuentas/src/main/java/com/banco/mscuentas/dto/MovimientoResponseDTO.java
@@ -2,6 +2,7 @@ package com.banco.mscuentas.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -14,7 +15,7 @@ public class MovimientoResponseDTO {
     private Long id;
     private LocalDateTime fecha;
     private String tipoMovimiento;
-    private Double valor;
-    private Double saldo;
+    private BigDecimal valor;
+    private BigDecimal saldo;
     private String numeroCuenta;
 }
